Update CMakeLists.txt and replace std::cerr with Logger in Array2D.h
In this commit, we have updated the CMakeLists.txt file to require C++23 standard and added MSVC compile options. Also, we replaced the usage of std::cerr in Array2D.h with a more suitable logger for debugging. This improves the readability and manageability of debug messages. Furthermore, the spelling error in the commented out 'propagate_const' has been corrected.