Add line comment ignoring functionality in Image.h
A new file IgnoreLineCommentsBuf.h was added containing classes that handle the ignoring of comment lines in the code. This feature was integrated into the Image.h file where it is now possible to ignore line comments in the PNM_header and read the desired image data efficiently. This change also results in lines of code being removed that were in the original get_image_format and read functions.